# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ZVideo API
  version: 1.0.0
  description: REST API for a video hosting platform.
servers:
  - url: /api/v1
security:
  - BearerAuth: []
...
paths:
...
  /videos/{videoId}/comments:
    get:
      tags: [Comments]
      security: []
      summary: List video comments
      operationId: listVideoComments
      parameters:
        - $ref: '#/components/parameters/VideoId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Comments page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentPage'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Comments]
      summary: Create video comment
      operationId: createVideoComment
      parameters:
        - $ref: '#/components/parameters/VideoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentRequest'
      responses:
        '201':
          description: Comment created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /comments/{commentId}:
    get:
      tags: [Comments]
      security: []
      summary: Get comment
      operationId: getComment
      parameters:
        - $ref: '#/components/parameters/CommentId'
      responses:
        '200':
          description: Comment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Comments]
      summary: Update comment
      operationId: updateComment
      parameters:
        - $ref: '#/components/parameters/CommentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentRequest'
      responses:
        '200':
          description: Comment updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Comments]
      summary: Delete comment
      description: Comment author, moderator, or admin can delete a comment.
      operationId: deleteComment
      parameters:
        - $ref: '#/components/parameters/CommentId'
      responses:
        '204':
          description: Comment deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /comments/{commentId}/rating:
    put:
      tags: [Comments]
      summary: Replace current user's comment rating
      operationId: rateComment
      parameters:
        - $ref: '#/components/parameters/CommentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingRequest'
      responses:
        '204':
          description: Rating applied.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /comments/{commentId}/stats:
    get:
      tags: [Comments]
      security: []
      summary: Get comment rating stats
      operationId: getCommentStats
      parameters:
        - $ref: '#/components/parameters/CommentId'
      responses:
        '200':
          description: Comment stats.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatingStats'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /channels/{channelId}/community:
    get:
      tags: [Community]
      security: []
      summary: Get channel community with posts
      operationId: getChannelCommunity
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Channel community.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/me/community:
    get:
      tags: [Community]
      summary: Get current user's channel community
      operationId: getMyCommunity
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Current user's community.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /channels/{channelId}/community/posts:
    post:
      tags: [Community]
      summary: Create community post
      operationId: createCommunityPost
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentRequest'
      responses:
        '201':
          description: Post created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPost'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /community/posts/{postId}:
    patch:
      tags: [Community]
      summary: Update community post
      operationId: updateCommunityPost
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentRequest'
      responses:
        '200':
          description: Post updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPost'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Community]
      summary: Delete community post
      operationId: deleteCommunityPost
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '204':
          description: Post deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /community/posts/{postId}/comments:
    get:
      tags: [Community]
      security: []
      summary: List community post comments
      operationId: listCommunityPostComments
      parameters:
        - $ref: '#/components/parameters/PostId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Community comments page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityCommentPage'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Community]
      summary: Create community post comment
      operationId: createCommunityComment
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentRequest'
      responses:
        '201':
          description: Community comment created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityComment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /community/comments/{commentId}:
    patch:
      tags: [Community]
      summary: Update community comment
      operationId: updateCommunityComment
      parameters:
        - $ref: '#/components/parameters/CommunityCommentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentRequest'
      responses:
        '200':
          description: Community comment updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityComment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Community]
      summary: Delete community comment
      operationId: deleteCommunityComment
      parameters:
        - $ref: '#/components/parameters/CommunityCommentId'
      responses:
        '204':
          description: Community comment deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ChannelId:
      name: channelId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    VideoId:
      name: videoId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    CommentId:
      name: commentId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    CommunityCommentId:
      name: commentId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Request validation or domain rule failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidUsername:
              summary: Invalid username
              value:
                error:
                  code: INVALID_USERNAME
                  message: The username is invalid
            invalidUserEmail:
              summary: Invalid email
              value:
                error:
                  code: INVALID_USER_EMAIL
                  message: The email is invalid
            weakPassword:
              summary: Weak password
              value:
                error:
                  code: WEAK_PASSWORD
                  message: The password is too weak
            playlistNameEmpty:
              summary: Empty playlist name
              value:
                error:
                  code: PLAYLIST_NAME_EMPTY
                  message: Playlist name cannot be empty
            playlistVideoChannelMismatch:
              summary: Playlist and video belong to different channels
              value:
                error:
                  code: PLAYLIST_VIDEO_CHANNEL_MISMATCH
                  message: Video belongs to another channel
            selfSubscription:
              summary: User cannot subscribe to own channel
              value:
                error:
                  code: SELF_SUBSCRIPTION
                  message: Cannot subscribe to your own channel
            invalidNotificationSettings:
              summary: Invalid notification settings
              value:
                error:
                  code: INVALID_NOTIFICATION_SETTINGS
                  message: Notification settings payload is invalid
            roleNotFound:
              summary: Role not found
              value:
                error:
                  code: ROLE_NOT_FOUND
                  message: Role not found
            communityPostContentEmpty:
              summary: Empty community post
              value:
                error:
                  code: COMMUNITY_POST_CONTENT_EMPTY
                  message: Community post content cannot be empty
            communityCommentContentEmpty:
              summary: Empty community comment
              value:
                error:
                  code: COMMUNITY_COMMENT_CONTENT_EMPTY
                  message: Community comment content cannot be empty
    Unauthorized:
      description: Authentication failed or token is missing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidCredentials:
              summary: Invalid login credentials
              value:
                error:
                  code: INVALID_CREDENTIALS
                  message: Invalid email or password
            invalidToken:
              summary: Invalid or expired token
              value:
                error:
                  code: INVALID_TOKEN
                  message: Invalid or expired token
            missingAuthorization:
              summary: Missing bearer token
              value:
                error:
                  code: UNAUTHORIZED
                  message: user not authorized
    Forbidden:
      description: Authenticated user is not allowed to perform this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              summary: Insufficient permissions
              value:
                error:
                  code: FORBIDDEN
                  message: You don't have permission to access this resource
            userBanned:
              summary: User account is banned
              value:
                error:
                  code: USER_BANNED
                  message: User account is banned
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            userNotFound:
              summary: User not found
              value:
                error:
                  code: USER_NOT_FOUND
                  message: User not found
            channelNotFound:
              summary: Channel not found
              value:
                error:
                  code: CHANNEL_NOT_FOUND
                  message: Channel not found
            videoNotFound:
              summary: Video not found
              value:
                error:
                  code: VIDEO_NOT_FOUND
                  message: Video not found
            playlistNotFound:
              summary: Playlist not found
              value:
                error:
                  code: PLAYLIST_NOT_FOUND
                  message: Playlist not found
            ratingNotFound:
              summary: Rating not found
              value:
                error:
                  code: RATING_NOT_FOUND
                  message: Rating not found
            commentNotFound:
              summary: Comment not found
              value:
                error:
                  code: COMMENT_NOT_FOUND
                  message: Comment not found
            communityPostNotFound:
              summary: Community post not found
              value:
                error:
                  code: COMMUNITY_POST_NOT_FOUND
                  message: Community post not found
            communityCommentNotFound:
              summary: Community comment not found
              value:
                error:
                  code: COMMUNITY_COMMENT_NOT_FOUND
                  message: Community comment not found
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: ERROR_CODE
        message:
          type: string
          example: Human-readable error message
...
    Channel:
      type: object
      required: [id, userId, ownerUsername, name, description, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        ownerUsername:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
...
    RatingAction:
      type: string
      enum: [like, dislike, remove]
    RatingRequest:
      type: object
      required: [action]
      properties:
        action:
          $ref: '#/components/schemas/RatingAction'
    RatingStats:
      type: object
      required: [likes, dislikes]
      properties:
        likes:
          type: integer
          minimum: 0
        dislikes:
          type: integer
          minimum: 0
    Comment:
      type: object
      required: [id, userId, username, videoId, content, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        videoId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    ContentRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
...
    Community:
      type: object
      required: [channel, posts, totalCount, limit, offset]
      properties:
        channel:
          $ref: '#/components/schemas/Channel'
        posts:
          type: array
          items:
            $ref: '#/components/schemas/CommunityPost'
        totalCount:
          type: integer
          format: int64
          minimum: 0
        limit:
          type: integer
        offset:
          type: integer
    CommunityPost:
      type: object
      required: [id, channelId, userId, username, content, createdAt]
      properties:
        id:
          type: integer
        channelId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CommunityComment:
      type: object
      required: [id, postId, userId, username, content, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    PageMeta:
      type: object
      required: [limit, offset, totalCount]
      properties:
        limit:
          type: integer
        offset:
          type: integer
        totalCount:
          type: integer
          format: int64
          minimum: 0
...
    CommentPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
...
    CommunityCommentPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/CommunityComment'